Create an empty "Today" object if no LogDays were loaded
Prevents a crash on startup. A TableView must have at least one section, however a fresh install has none. Creating a fake section without an underlying LogDay causes a problem when the first LogDay is created and its corresponding insertion event is generated. UITableView sees the event and expects its section count to increase, which doesn't happen because the number of displayed sections hasn't changed, so an exception is thrown and the app is terminated.