Add selectedInsulinTypes property to InsulinTypesViewController
And remove the editedObject and editedIndex properties. The view controller shouldn't need to know anything about the object that's being edited; it should only care about presenting options to the user for selection. The selectedInsulinTypes property exists so the parent can set the initially selected types when editing an existing record and for selecting the new entry defaults.