Fix adding of new Categories and Insulin Types
Wasn't scrolling to the new row and wasn't setting edit mode on the new row. 

Insert the new row and then check to see if it's visibile. If it is, make it the first responder. If not, use performSelector to try again after a delay (to wait for any scrolling to finish).

This is a nasty kludge that could potentially go into an infinite loop. Fortunately it generally doesn't.